Issue #966 - fixes problem with ConnectionIterator
If the iterator contained both cursors and the page urls are added by
code (and not via the original JSON) it was not possible to stop the
iteration. The problem was the before and after cursor. In the special
case these values are equal and so a endless loop was generated.

Luckily it was finished by a NPE, so it was not endless at all. But it
is fixed by checking the cursor values. Are before and after the same we
think there is no other page and stop the pagination.

Please check the json in case of a unusual behavior and open an issue.